Add "flat" drill and organizational folders for sprites
I would like to keep an "unused" sprites folder in case anyone else wants to use them and a "new" folder just so it's easier to organize and add/update new sprites. The license is also just for giving permission to anyone to reuse the sprites in their own mods. The new drill is 3x3, named 1-9 for each cell starting at the top left, feel free to rename it "quarry drill" or something if you don't want to use flat drill.

Positioning reference:
1 2 3
4 5 6
7 8 9